+++ NEWS.md
@@ -19,6 +19,9 @@
 * mkvmerge, GUI's multiplexer: MIME types: added the `font` top-level media
   types from RFC 8081. This means that the following new MIME types for fonts
   can be used: `font/ttf`, `font/otf`, `font/woff` and `font/woff2`.
+* mkvmerge: fixed reading text files encoded in UTF-16 oder UTF-32 that have
+  different forms of line endings (new lines, carriage returns or a mix of
+  both). Fixes #2160.
 
 
 # Version 18.0.0 "Apricity" 2017-11-18